+++ README.md
@@ -10,4 +10,4 @@
 1. follow steps from [Set up DEV in your account](#set-up-dev-in-your-account)
 2. `export REPORT_TABLE_NAME=$(aws ssm get-parameter --name "/core/DynamoDbStack/Tables/Report" --query "Parameter.Value" --output text)`
 3. `npx projen express:run`
-4. [test your api](src/stacks/api/report/debug/report-api.http)
+4. [test your api](src/stacks/api-gateway/api/debug/report-api.http)
