+++ src/stacks/api-gateway/api/debug/report-api.http
