+++ src/stacks/api-gateway/api/express-app/config/dynamoDB.ts
