+++ src/stacks/api-gateway/api/express-app/interfaces.ts
