+++ src/stacks/api-gateway/api/express-app/middlewares/errorMiddleware.ts
