+++ src/stacks/api-gateway/api/express-app/middlewares/loggerMiddleware.ts
